+++ src/mdacli/__init__.py
@@ -11,4 +11,4 @@ from mdacli.cli import cli
 
 __all__ = ["cli"]
 
-__version__ = '0.1.11'
+__version__ = '0.1.12'